requirements.txt
fastapi
uvicorn
python-multipart
requests
pypdf
langchain
langchain-openai
langchain-community
python-dotenv